Refactor based on PR comments + remove some unused code
